tdf#157696 a11y: Map "CharEscapement" text attr to IA2 "text-position"
While the gtk3 implementation (see `Escapement2VerticalAlign`
in `vcl/unx/gtk3/a11y/atktextattributes.cxx`) bridges numeric
values to ATK/AT-SPI, the IAccessible2 text attribute specification [1]
only allows the 3 values "baseline", "super" and "sub".

With this in place, Orca with the qt6 VCL plugin on Linux
and NVDA on Windows (with the WIP branches to switch both
LO's winaccessibility and NVDA to the IAccessible2 text
attributes) announce the subscript and superscript text
in the tdf#157696 sample document as expected.
(For Orca, announcement of the "vertical-align" text
attribute is off by default, so that first needs to be enabled
in the "Text Attributes" section of the Orca configuration
dialog.)

[1] http://www.w3.org/TR/REC-CSS2/visudet.html#propdef-vertical-align

Change-Id: I1e4ea014cf818fefd009089be183615abf67aada
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/157867
Tested-by: Jenkins
Reviewed-by: Michael Weghorn <m.weghorn@posteo.de>