sw floattable, wrap on all pages: add layout
The anchor text of a floating table is normally wrapped around the table
only on the last page of the table. This compat flag requests to wrap on
all pages instead.

First, disable the SwTextFrame::FormatEmpty() optimization in this case,
so not only the last page has portions inside the text frame.

Second, improve SwTextFrame::FormatAdjust(), so that anchor content from
the last page can flow to previous pages.

Note that having a dedicated SwFlyAtContentFrame::IsWrapOnAllPages()
means that later it'll be possible to do per-frame decision on this if
there is demand for that (this came up on the mailing list, but it's not
something Word would support).

Change-Id: I75df539df859aaa6c7eaaeddb33f004639b1c004
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/157902
Reviewed-by: Miklos Vajna <vmiklos@collabora.com>
Tested-by: Jenkins