tdf#157696 a11y Translate "CharStrikeout" to IA2 equivalents
The IAccessible2 text attributes specification [1]
has several different "text-line-through-*"
attributes for indicating strikethrough.

With this change in place, NVDA on Windows (with the WIP branches
to switch both LO's winaccessibility and NVDA to the
IAccessible2 text attributes) announces the striked out
text in the tdf#157696 sample document as expected.

For the qt6 VCL plugin on Linux, this is not the
case yet, because Qt's AT-SPI adaptor does not support
the attributes yet; s.a. QTBUG-118106 [2].

Pending qtbase change implementing that: [3]

While the gtk3 VCL plugin reports 2 attributes, e.g.
"strikethrough: single" and "text-decoration: line-through",
the ATK text attribute specification [4] only mentions a
`ATK_TEXT_ATTR_STRIKETHROUGH` that can be either
"true" or "false", and that is also what e.g.
the gtk4-demo "Text View" -> "Markup" example does,
so the suggested Qt change also does this, which means
that just "strikethrough: true" will be reported when
any kind of strikethrough is in place with the qt6 VCL
plugin and both, this LibreOffice change and the
pending Qt Gerrit change applied.

[1] https://wiki.linuxfoundation.org/accessibility/iaccessible2/textattributes
[2] https://bugreports.qt.io/browse/QTBUG-118106
[3] https://codereview.qt-project.org/c/qt/qtbase/+/511758
[4] https://gnome.pages.gitlab.gnome.org/atk/AtkText.html#AtkTextAttribute

Change-Id: Ib4ebba183be233aacfcf33a4fff29b27711d2e1e
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/157939
Tested-by: Jenkins
Reviewed-by: Michael Weghorn <m.weghorn@posteo.de>