qtcreator: Specify C++ std version using CONFIG qmake var
Extract the C++ standard version to use that is specified
by a `-std=<version>` or `-std:<version>` compiler flag,
and set that via "CONFIG += <version>" in the .pro files
used by Qt Creator.

This makes the Clang Code Model use the correct mode
and no longer complain about `char8_t` after the
switch to C++20 in my Windows development setup
in Qt Creator:

> accessibletabbarpagelist.hxx:22:10: In included file: use of undeclared identifier 'char8_t'
> stringutils.hxx:252:31: error occurred here

The previous way of specifying this via
`QMAKE_CXXFLAGS` as introduced in

    commit 92c03d9bf644b0f10de52ce0da09f97056e46247
    Author: Michael Weghorn <m.weghorn@posteo.de>
    Date:   Fri Jun 7 21:44:03 2019 +0200

        qtcreator: Take over '-std=...' from CXXFLAGS

appears to not (no longer?) work at least with Qt Creator 11.0.3
on Windows.

On the contrary however, setting only `CONFIG` and
not `QMAKE_CXXFLAGS` causes the exact same problem
in my Linux setup.

Therefore, set both qmake variables.

Note that there is a specific set of accepted values
for the the `CONFIG` variable in .pro files [1],
but at least "c++20" and "c++latest" are accepted
and have the expected meaning, so that should be
fine for now.

[1] https://doc.qt.io/qt-6/qmake-variable-reference.html#config

Change-Id: Idc75b74300c7bdd0f6193fcfc1758b536728b887
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/158053
Tested-by: Jenkins
Reviewed-by: Michael Weghorn <m.weghorn@posteo.de>