tdf#157788 Saving the "Show navigation panel" setting for a device
Currently, the value of the "Show navigation panel" parameter for
presentations is not saved. The section in which this setting is
now located, apparently, is saved in files. It might make sense
to move this setting to Tools -> Options -> LibreOffice Impress -> General,
because it relates to the device rather than the slideshow.

Change-Id: I2286a4a6d432b11fce2b9c3e65fa6b82e004275f
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/158057
Tested-by: Jenkins
Reviewed-by: Heiko Tietze <heiko.tietze@documentfoundation.org>