tdf#148725: Reset must clear respective items from m_pOutSet
After Reset, the tabbed dialog's output set should restore to the same
state (concerning the current tab page) as immediately after creation.
This state means that pressing OK immediately afer Reset would not try
to set or remove any items in the target. SfxTabDialogController::Ok
has this code:

    if (m_pOutSet && m_pOutSet->Count() > 0)
        bModified = true;

meaning that m_pOutSet is expected to be empty in case of completely
unmodified dialog; and Reset must make sure this for a given page.

Instead, commit 28fc0962b10519ab84654d189d2ad0cca8f84f95 (weld SwLabDlg,
2018-04-27) made Reset handler to populate the output item set with all
the items from the input set, that belong to the page's which ranges.
This made all the settings set in parents (which therefore appeared in
the input set) to be set directly in the target upon application.

Change-Id: Iacfffb5670cc3ade950ac412b818acb482845255
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/158067
Tested-by: Jenkins
Reviewed-by: Mike Kaganski <mike.kaganski@collabora.com>