tdf#157026 sd notebookbar: add ObjectAlign to MenuTable as popupmenu
The vast majority of this is just eliminating copy/paste,
since I have now created a popupmenu.xml list for the popup menu.

The "fix" is adding this to the Table menus.

.uno:ObjectAlign is basically just a "label".
Now it is a label for two different "views":
* a toolbar widget dropdown
* a menu subcategory

Perhaps it is confusing to have one UNO entry
serve these two similar purposes?

I suppose someone might also complain that a simple
    <menu:menu menu:id=".uno:ObjectAlign"/>
isn't clear that this is a submenu.

I considered making mine a new ObjectAlignMenu,
but the annoyance of having to create a duplicate string translation in
officecfg/registry/data/org/openoffice/Office/UI/GenericCommands.xcu
made me just reuse the existing label.

Change-Id: I17e037e0c9c5d1f57ac9f174a76cd041983bcea4
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/157894
Tested-by: Jenkins
Reviewed-by: Justin Luth <jluth@mail.com>