Do not forget to notify in cases when modified redline's start/end don't move
SwRangeRedline::SetStart/SetEnd call MaybeNotifyRedlineModification,
which means that e.g. continuing typing at the end of the current
redline would continue sending notifications. But as soon as cursor
is moved (e.g. using arrow left, inside the current redline), the
next modification is completely inside the current one, and would
be simply deleted. In this case (and other similar), notifications
weren't sent.

This change tries to make sure that every such case is handled. In
case of delete-inside-insert, when positions are equal, the current
redline is deleted; MaybeNotifyRedlineModification is not called,
to not crash. I don't know if this needs additional notification.

Change-Id: I2a97354623a5a784928ce4e1f2c93606954bd11c
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/158198
Tested-by: Jenkins
Reviewed-by: Mike Kaganski <mike.kaganski@collabora.com>