use Liberation Sans instead of Calibri in the test for tdf#144092
based on Stephan Bergmann's point on:
<https://gerrit.libreoffice.org/c/core/+/138652/comment/bb5fc1a4_75e060c9/>

the font used in the test file was Calibri, which likely was
being replaced with the best approximation on a given system.
Changed that to Liberation Sans which is a font that is bundled
with LibreOffice to improve the test's robustness.

Change-Id: I7ff75baeb9259dea244913ca9d5025948291f1e2
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/158234
Reviewed-by: Stephan Bergmann <sbergman@redhat.com>
Tested-by: Jenkins
(cherry picked from commit 676e0527d2f31556eccae314fbb12ce204f02ec7)
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/158277
Reviewed-by: Xisco Fauli <xiscofauli@libreoffice.org>