tdf#135922 a11y: Report IA2 "text-align" attr
Map the "ParaAlign" UNO text attribute to the
IAccessibl2 *object* attribute "text-align",
as specified in the IAccessible2 object
attributes specification. [1]

This causes the corresponding "justification" AT-SPI text
attribute to be reported when using the qt6 VCL
plugin on Linux.

For Windows, bridging this to the platform a11y
will be implemented in an upcoming change.

See the commmit message of the previous change

    Change-Id Ief7c840d3c5274714a914ca0e56df0c5eaffb06d
    tdf#135922 a11y: Prepare reporting text attrs as IA2 obj attrs

for more details.

[1] https://wiki.linuxfoundation.org/accessibility/iaccessible2/objectattributes

Change-Id: I6a0478a4ac37f69393e995602ed698b5a78488e5
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/158256
Tested-by: Jenkins
Reviewed-by: Michael Weghorn <m.weghorn@posteo.de>