a11y: Improve XAccessibleExtendedAttributes doc
The existing

    // MT: I guess it's not formula only?

comment was justified.

`XAccessibleExtendedAttributes::getExtendedAttributes`,
is not only used to report formulae, but e.g. used to
report a heading level by Writer paragraphs
(s. `SwAccessibleParagraph::getExtendedAttributes`),
so adjust the documentation and align it with
the `IAcccessible2::attributes` one [1] that also
uses the cell formula as an example.

[1] https://accessibility.linuxfoundation.org/a11yspecs/ia2/docs/html/interface_i_accessible2.html#a0d51b0c189a000ee3b6ddf7f68da2009

Change-Id: I2d6f38911c3802edb55f9426a35b8986dd6e399a
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/158259
Tested-by: Jenkins
Reviewed-by: Michael Weghorn <m.weghorn@posteo.de>