tdf#135922 wina11y: Report text-related IA2 obj attrs
Extend `CMAccessible::get_attributes` so that
it also reports the text-related IAccessible2
object attributes, since what is meant to be
reported as a text attribute and what is meant
to be reported as an object attribute differs
between the IAccessible2 specificiation and what
LibreOffice does on the UNO level, s. the commit message
in this previous change for more details:

    Change-Id Ief7c840d3c5274714a914ca0e56df0c5eaffb06d
    tdf#135922 a11y: Prepare reporting text attrs as IA2 obj attrs

Just use a character offset of 0 when querying via the
`XAccessibleText` interface here. The exct offset used
used shouldn't make any difference for paragraph-specific
attributes.

With this and the NVDA pull request [1] to evaluate attributes
according to the IAccessible2 text attributes and
IAccessible2 object attributes specifications, NVDA
now reports the alignment of paragraphs in Writer, e.g.
says "align center" since the corresponding attribute is
now reported for the paragraph object, as can also
be seen by querying the IAccessible2 interface manually
in NVDA's Python console:

    >>> focus.IAccessibleObject.attributes
    'heading-level:;level:;text-align:center;'

("text-align:center;" was not yet reported without
this change in place.)

[1] https://github.com/nvaccess/nvda/pull/15649

Change-Id: I5723797232f89db6a2b74d4a601344f2078ee630
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/158260
Tested-by: Jenkins
Reviewed-by: Michael Weghorn <m.weghorn@posteo.de>