tdf#157817 sw: PDF/UA export: fix ToX header sections
The problem was that a P or equivalent element is not allowed as child
of a TOC; previously the P was outside/before the TOC because the ToX
header was completely ignored but now it starts a SE for the ToC content
section before being ignored.

Tagged PDF Best Practice Guide suggests putting the header inside a
Caption element.

Another problem is that if the ToX header is on one page but the first
entry in the ToX content section is on the next page, BeginTag() created
a new SE for the ToX content section.

(regression from commit 033e37b49b5712df19dbfd2c307a102bce620de1)

Change-Id: I3c1f8fdda81c1f10198f5eaefc10bcd14f277a8b
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/158298
Tested-by: Jenkins
Reviewed-by: Michael Stahl <michael.stahl@allotropia.de>
(cherry picked from commit a5851d0a6fdc3e08e93daf8cf0ae7f6fc2d73ef6)
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/158285
Reviewed-by: Miklos Vajna <vmiklos@collabora.com>