SW: small fix for manage changes dialog.
Fixed a continue, as it was planned to effect the outer loop.
(probably when it was replaced to an inner loop it was forgot to handle)

fixed a nullptr check also.

These problems may not result real problem right now,
but when interdependedt redlines will be improved,
they will result in problems.

Change-Id: Ie139f9a657abbacea9450176d2fe138392beddfa
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/158457
Tested-by: Jenkins CollaboraOffice <jenkinscollaboraoffice@gmail.com>
Reviewed-by: Caolán McNamara <caolan.mcnamara@collabora.com>
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/158613
Tested-by: Caolán McNamara <caolan.mcnamara@collabora.com>