tdf#156988 chardlg: default to AUTO escapement
As of 24.2, the default value for a superscript or subscript
is automatic placement since d384ccdb04ebeb8b094e6d9a2ddf4e5aea5327c8.

Therefore, it makes sense for the character properties
dialog to also default to automatic.

Normally this happens - since normally there is some kind of context,
and in that case m_xNormalPosBtn was set with AUTO (m_xHighLowRB).
However, in the case of finding a formating, there is no context
for the very first run. That is what this patch intended to fix,
since in that case too we want m_xHighLowRB to be set to true.

For any other similar situation where there is no context,
assuming DFLT_ESC_AUTO_SUPER or DFLT_ESC_AUTO_SUB
should be appropriate.

Change-Id: I1ee78c72bbb1c3d23fa39da29322436632f12b14
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/158514
Tested-by: Jenkins
Reviewed-by: Justin Luth <jluth@mail.com>