tdf#135586 wina11y: Make role mapping more explicit
The previous `ROLE_TABLE` looked like it might
explicitly map the UNO role on the left-hand side
to the IAccessible2 role on the right-hand side,
and the comment also said so.

In reality, the UNO role on the left hand side was
not really accessed, but the assumption was that
the index of each entry would correspond to the
sal_Int16 value of the UNO role
(`css::accessibility::AccessibleRole`).

Make that mapping more explicit by replacing
it with a helper function
`lcl_mapToIAccessible2Role` and slightly
simplify `AccObject::UpdateRole`.

Change-Id: Id02553c4df042982f92fce3338077e95db687a83
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/158572
Tested-by: Jenkins
Reviewed-by: Michael Weghorn <m.weghorn@posteo.de>