Resolves: tdf#157891 arrange so 'ok' in paste special is default action
for return.

Set the "all" checkbox explicitly as the widget to have the initial
focus, all platforms then agree on where focus starts, and return will
activate the default "ok" button.

Otherwise, vcl picks the "value" button and gtk picks the "link"
checkbox and there is an inconsistency as to what "return" activates by
default:

Change-Id: I894199069a32f4e09e2661f2c90338caacb8c65d
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/158569
Tested-by: Jenkins
Reviewed-by: Caolán McNamara <caolan.mcnamara@collabora.com>