tdf#135586 sw a11y: Use BLOCK_QUOTE role for "Block Quotation" para
If a paragraph is using the "Block Quotation" paragraph style,
report the newly introduced BLOCK_QUOTE a11y role for it instead
of the PARAGRAPH role.

This is similar to how the HEADING role is reported when
one of the paragraph styles for headings is used
(s. `SwAccessibleParagraph::GetRealHeadingLevel`).

This is also in line with what is documented for mapping
LO elements to tagged PDF elements in
sw/inc/EnhancedPDFExportHelper.hxx:

> * Mapping of OOo elements to tagged pdf elements:
> *
> * OOo element                              tagged pdf element
> * -----------                              ------------------
> * [...]
> * SwFormat Quotations                       BlockQuote

This makes the Orca screen reader with the gtk3 VCL plugin on Linux
and the NVDA screen reader on Windows explicitly announce a paragraph
as block quote when moving there with the cursor.

Adapt some places that have specific handling for the
PARAGRAPH role to take into account the BLOCK_QUOTE
role as well.

Change-Id: I4a89625c2a3f07d37df09e68cb7045a59cfff633
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/158574
Tested-by: Jenkins
Reviewed-by: Michael Weghorn <m.weghorn@posteo.de>