qtcreator: Add Obj-C++ files to project
Extend gbuild-to-ide to also take into account
Objective-C++ files that are set with the
`OBJCXXOBJECTS` key in the json files.

Extend the Qt Creator IDE integration target
(`make qtcreator-ide-integration`) to make use
of that information to find the corresponding
headers as well, and set sources and headers
in the `OBJECTIVE_SOURCES` [1] and
`OBJECTIVE_HEADERS` [2] qmake veriables.

This way, those files are part of the project
and the Clang Code Model works as expected,
which makes features like code completion and
showing issues work correctly for such files
(like e.g. vcl/osx/a11ywrapper.mm) as well.

[1] https://doc.qt.io/qt-6/qmake-variable-reference.html#objective-sources
[2] https://doc.qt.io/qt-6/qmake-variable-reference.html#objective-headers

Change-Id: I2c1cf5bdfdc64300ec3eb77e42de1ae2cd4223f4
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/158665
Tested-by: Jenkins
Reviewed-by: Michael Weghorn <m.weghorn@posteo.de>