macOS a11y: Don't report (non-)expanded state when not expandable
When the a11y object is not expandable, also don't
report a true/false value for the expanded state on
macOS.

Otherwise, the CrossOver screen reader on macOS
will explicitly announce "collapsed" for these
objects, e.g. when moving focus between paragraphs
in Writer (as reported on the a11y mailing list [1])
or between toolbar butons, since
`isAccessibilityExpanded` gets called and
returns false, so it is assumed that this
is an expandable item that is currently
not expanded, i.e. it is collapsed.

Quoting from the NSAccessibilityProtocol doc [2]:

> You can control which accessor methods the assistive app can
> use by overriding isAccessibilitySelectorAllowed(_:).
> Return true if the assistive app can call the selector;
> otherwise, return false."

[1] https://listarchives.libreoffice.org/global/accessibility/msg01068.html
[2] https://developer.apple.com/documentation/appkit/nsaccessibilityprotocol:

Change-Id: Iac9027a598e5793d6bd0da79212d7b07448a4c34
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/158733
Tested-by: Jenkins
Reviewed-by: Patrick Luby <plubius@neooffice.org>
Reviewed-by: Michael Weghorn <m.weghorn@posteo.de>